fix: "🔨데이터가 여러번 로드 되는 문제"
Application레벨에서 onCreate가 될때 초기화 해서 해결 했다
